Update MailSendController.cs
Update of method: CreateGeneratethemailsendrequestAsync
When we don't provide the "url" parameter and the default case is null, the _queryBuilder is = "/mail/send". 
The method "APIHelper.CleanUrl(_queryBuilder)"  checks if the input (in this case _queryBuilder) starts with https ( Match match = Regex.Match(input, "^https?://[^/]+")). 
Because the default Configuration.BaseUri is not included in the _queryBuilder and the _queryBuilder doesn't start with "https" , I always get the error for Invalid Url format!